% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfs.R
\name{add_acfs.mcgf}
\alias{add_acfs.mcgf}
\title{Calculating mean autocorrelations for an \code{mcgf} object}
\usage{
\method{add_acfs}{mcgf}(x, lag_max, ...)
}
\arguments{
\item{x}{An \code{mcgf} object.}

\item{lag_max}{Maximum lag at which to calculate the acf.}

\item{...}{Additional parameters or attributes.}
}
\value{
A vector of mean autocorrelations.
}
\description{
Calculating mean autocorrelations for an \code{mcgf} object
}
\details{
It computes mean autocorrelations for each time lag across locations, and add
them to \code{x}.
}
\examples{
data <- cbind(S1 = 1:5, S2 = 4:8, S3 = 5:9)
lon <- c(110, 120, 130)
lat <- c(50, 55, 60)
locations <- cbind(lon, lat)
obj <- mcgf(data = data, locations = locations)
obj <- add_acfs(x = obj, lag_max = 2)
print(obj, "acfs")
}
\seealso{
Other {functions related to the class}: 
\code{\link{add_acfs}()},
\code{\link{dists.mcgf}()},
\code{\link{dists}()},
\code{\link{mcgf}()},
\code{\link{print.mcgf}()}
}
\concept{{functions related to the class}}
