% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_base.R
\name{add_base}
\alias{add_base}
\title{Generic function for adding a base model}
\usage{
add_base(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
\code{x} with the newly added base model.
}
\description{
Generic function for adding a base model
}
\details{
Refer to \code{\link[=add_base.mcgf]{add_base.mcgf()}} and \code{\link[=add_base.mcgf_rs]{add_base.mcgf_rs()}} for more details.
}
