% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfs.R, R/sds.R
\name{add_ccfs.mcgf}
\alias{add_ccfs.mcgf}
\alias{`sds<-`}
\alias{sds<-}
\title{Calculating cross-correlations for an \code{mcgf} object}
\usage{
\method{add_ccfs}{mcgf}(x, lag_max, ...)

sds(x) <- value
}
\arguments{
\item{x}{An \code{mcgf} object.}

\item{lag_max}{Maximum lag at which to calculate the acf.}

\item{...}{Additional parameters or attributes.}

\item{value}{A vector of standard deviations for all stations.}
}
\value{
An \code{mcgf} object with newly added cross-correlations and a vector
of standard deviations under attributes \code{ccfs} and \code{sds} respectively.
}
\description{
Calculating cross-correlations for an \code{mcgf} object
}
\details{
It computes cross-correlations for each time lag, and add them to \code{x}.
The output is an array of matrices where each matrix corresponds to the
cross-correlation for a time lag, and a vector of empirical standard
deviations. If other standard deviation functions are used, they can be
replaced via \link{sds<-}.
}
\examples{
wind_sq <- sqrt(wind[, -1])
time <- wind[, 1]
wind_mcgf <- mcgf(data = wind_sq, locations = wind_loc, time = time)
wind_mcgf <- add_ccfs(x = wind_mcgf, lag_max = 3)
print(wind_mcgf, "ccfs")
print(wind_mcgf, "sds")

}
\seealso{
Other {functions related to the auto- and cross-correlations}: 
\code{\link{acfs.mcgf}()},
\code{\link{acfs}()},
\code{\link{add_acfs.mcgf}()},
\code{\link{add_acfs}()},
\code{\link{add_ccfs}()},
\code{\link{ccfs.mcgf}()},
\code{\link{ccfs}()}
}
\concept{{functions related to the auto- and cross-correlations}}
