% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lagr.R
\name{add_lagr}
\alias{add_lagr}
\title{Generic function for adding a Lagrangian model}
\usage{
add_lagr(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
\code{x} with the newly added Lagrangian model.
}
\description{
Generic function for adding a Lagrangian model
}
\details{
Refer to \code{\link[=add_lagr.mcgf]{add_lagr.mcgf()}} and \code{\link[=add_lagr.mcgf_rs]{add_lagr.mcgf_rs()}} for more details.
}
