% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfs.R
\name{ccfs}
\alias{ccfs}
\title{Generic function for calculating cross-correlation}
\usage{
ccfs(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
An array of cross-correlations for \code{mcgf} objects, or that plus
a list of regime-switching cross-correlations for \code{mcgf_rs} objects.
}
\description{
Generic function for calculating cross-correlation
}
\details{
Refer to \code{\link[=ccfs.mcgf]{ccfs.mcgf()}} and \code{\link[=ccfs.mcgf_rs]{ccfs.mcgf_rs()}} for more details.
}
