% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccov.R
\name{ccov}
\alias{ccov}
\title{Generic functions for calculating joint covariance/correlation matrix for mcgf
objects}
\usage{
ccov(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
Joint correlation/covariance matrix.
}
\description{
Generic functions for calculating joint covariance/correlation matrix for mcgf
objects
}
\details{
Refer to \code{\link[=ccov.mcgf]{ccov.mcgf()}} and \code{\link[=ccov.mcgf_rs]{ccov.mcgf_rs()}} for more details.
}
