% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dist_sign}
\alias{check_dist_sign}
\title{Check if valid signed distance}
\usage{
check_dist_sign(x, name)
}
\arguments{
\item{x}{Distance matrix or array.}
}
\value{
NULL.
}
\description{
Check if valid signed distance
}
\details{
Check if \code{x} is a valid signed distance vector, matrix or array. It errors
if \code{x} in absolute value is not a symmetric matrix or an array of
symmetric matrices.
}
\keyword{internal}
