% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dists}
\alias{check_dists}
\title{Check if valid dists attribute for an \code{mcgf} object}
\usage{
check_dists(dists, n_var, names, name_dists = "dists")
}
\arguments{
\item{dists}{List of scaler or vector}

\item{n_var}{Scalar, number of variables.}

\item{names}{column and row names of matrices in \code{dists}.}

\item{name_dists}{name_dists of \code{dists}.}
}
\value{
\code{dists}.
}
\description{
Check if valid dists attribute for an \code{mcgf} object
}
\details{
Check if \code{dists} is a valid dists attribute for an \code{mcgf} object. It errors
if 1) \code{dists} does not contain \code{h1} or \code{h2}, 2) if their dimensions do not
match, 3) if it contains elements other than \code{h}, \code{h1} or \code{h2}. \code{h} will be
computed if it is not given.
}
\keyword{internal}
