% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_fs.R
\name{cor_fs}
\alias{cor_fs}
\title{Calculate correlation for fully symmetric model}
\usage{
cor_fs(nugget = 0, c, gamma = 1/2, a, alpha, beta = 0, h, u)
}
\arguments{
\item{nugget}{The nugget effect \eqn{\in[0, 1]}.}

\item{c}{Scale parameter of \code{cor_exp}, \eqn{c>0}.}

\item{gamma}{Smooth parameter of \code{cor_exp}, \eqn{\gamma\in(0, 1/2]}.}

\item{a}{Scale parameter of \code{cor_cauchy}, \eqn{a>0}.}

\item{alpha}{Smooth parameter of \code{cor_cauchy}, \eqn{\alpha\in(0, 1]}.}

\item{beta}{Interaction parameter, \eqn{\beta\in[0, 1]}.}

\item{h}{Euclidean distance matrix or array.}

\item{u}{Time lag, same dimension as \code{h}.}
}
\value{
Correlations of the same dimension as \code{h} and \code{u}.
}
\description{
Calculate correlation for fully symmetric model
}
\details{
The fully symmetric correlation function with interaction parameter
\eqn{\beta} has the form
\deqn{C(\mathbf{h}, u)=\dfrac{1}{(a|u|^{2\alpha} + 1)}
\left((1-\text{nugget})\exp\left(\dfrac{-c\|\mathbf{h}\|^{2\gamma}}
{(a|u|^{2\alpha}+1)^{\beta\gamma}}\right)+
\text{nugget}\cdot \delta_{\mathbf{h}=\boldsymbol 0}\right),}
where \eqn{\|\cdot\|} is the Euclidean distance, and \eqn{\delta_{x=0}} is 1
when \eqn{x=0} and 0 otherwise. Here \eqn{\mathbf{h}\in\mathbb{R}^2} and
\eqn{u\in\mathbb{R}}. By default \code{beta = 0} and it reduces to the separable
model.
}
\examples{
h <- matrix(c(0, 5, 5, 0), nrow = 2)
u <- matrix(0, nrow = 2, ncol = 2)
cor_fs(c = 0.01, gamma = 0.5, a = 1, alpha = 0.5, beta = 0.5, h = h, u = u)

h <- array(c(0, 5, 5, 0), dim = c(2, 2, 3))
u <- array(rep(0:2, each = 4), dim = c(2, 2, 3))
cor_fs(c = 0.01, gamma = 0.5, a = 1, alpha = 0.5, beta = 0.5, h = h, u = u)

}
\references{
Gneiting, T. (2002). Nonseparable, Stationary Covariance Functions for
Space–Time Data, Journal of the American Statistical Association, 97:458,
590-600.
}
\seealso{
Other correlation functions:
\code{\link[=cor_cauchy]{cor_cauchy()}},
\code{\link[=cor_exp]{cor_exp()}},
\code{\link[=cor_lagr_askey]{cor_lagr_askey()}},
\code{\link[=cor_lagr_exp]{cor_lagr_exp()}},
\code{\link[=cor_lagr_tri]{cor_lagr_tri()}},
\code{\link[=cor_sep]{cor_sep()}},
\code{\link[=cor_stat]{cor_stat()}},
\code{\link[=cor_stat_rs]{cor_stat_rs()}}
}
\concept{correlation functions}
