% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dists}
\alias{dists}
\title{Generic function for calculating distance matrices}
\usage{
dists(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
A list of signed distance matrices: \code{h} (Euclidean), \code{h1}
(horizontal), and \code{h2} (vertical) with the same dimensions.
}
\description{
Generic function for calculating distance matrices
}
\concept{functions related to the class}
