% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_stat.R
\name{.corr_stat}
\alias{.corr_stat}
\title{Calculate general stationary correlation.}
\usage{
.corr_stat(
  par_base = NULL,
  par_lagr,
  base,
  lagrangian,
  lambda = lambda,
  base_fixed = FALSE
)
}
\arguments{
\item{par_base}{Parameters for the base model (symmetric), used only when
\code{base_fixed = FALSE}.}

\item{par_lagr}{Parameters for the Lagrangian model.}

\item{base}{Base model, \code{sep} or \code{fs} for now. Or correlation matrix/array.}

\item{lagrangian}{Lagrangian model, \code{lagr_tri} for now.}

\item{lambda}{Weight of the Lagrangian term, \eqn{\lambda\in[0, 1]}.}

\item{base_fixed}{Logical; if TRUE, \code{base} is the correlation.}
}
\value{
Correlations for the general stationary model. Same dimension of
\code{base} if \code{base_fixed = FALSE}.
}
\description{
Calculate general stationary correlation.
}
\details{
The general station model, a convex combination of a base model
and a Lagrangian model, has the form
\deqn{C(\mathbf{h}, u)=(1-\lambda)C_{\text{Base}}(\mathbf{h}, u)+
\lambda C_{\text{Lagr}}(\mathbf{h}, u),}
where \eqn{\lambda} is the weight of the Lagrangian term.

If \code{base_fixed = TRUE}, then the correlation is of the form
\deqn{C(\mathbf{h}, u)=(1-\lambda)\text{Base}+
\lambda C_{\text{Lagr}}(\mathbf{h}, u).}
}
\keyword{internal}
