% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Optimization for wls method}
\usage{
estimate(par_init, method, optim_fn, cor_fn, par_fixed, lower, upper, ...)
}
\arguments{
\item{par_init}{Initial values for parameters to be optimized.}

\item{method}{Parameter estimation method. "wls" or "mle".}

\item{optim_fn}{Optimization function.}

\item{cor_fn}{Correlation function.}

\item{par_fixed}{Fixed parameters of \code{cor_fn}.}

\item{lower}{Lower bound.}

\item{upper}{Upper bound.}

\item{...}{Additional arguments passed to \code{optim_fn}, \code{obj_wls} or \code{obj_mle}.}
}
\value{
A list outputted from optimization functions of \code{optim_fn}.
}
\description{
Optimization for wls method
}
\keyword{internal}
