% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_base.R
\name{fit_base}
\alias{fit_base}
\title{Fit correlation base models}
\usage{
fit_base(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
A vector of estimated parameters.
}
\description{
Fit correlation base models
}
\details{
Refer to \code{\link[=fit_base.mcgf]{fit_base.mcgf()}} and \code{\link[=fit_base.mcgf_rs]{fit_base.mcgf_rs()}} for more details.
}
