% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lagr.R
\name{fit_lagr}
\alias{fit_lagr}
\title{Fit correlation Lagrangian models}
\usage{
fit_lagr(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
A vector of estimated parameters.
}
\description{
Fit correlation Lagrangian models
}
\details{
Refer to \code{\link[=fit_lagr.mcgf]{fit_lagr.mcgf()}} and \code{\link[=fit_lagr.mcgf_rs]{fit_lagr.mcgf_rs()}} for more details.
}
