% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_numeric_scalar}
\alias{is_numeric_scalar}
\title{Check if numeric scalar}
\usage{
is_numeric_scalar(x)
}
\arguments{
\item{x}{Input}
}
\value{
Logical.
}
\description{
Check if numeric scalar
}
\details{
Check if \code{x} is a numeric scalar.
}
\keyword{internal}
