% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige.R
\name{krige}
\alias{krige}
\title{Generic function for computing kriging forecasts}
\usage{
krige(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
Kriging results of \code{x}.
}
\description{
Generic function for computing kriging forecasts
}
\details{
Refer to \code{\link[=krige.mcgf]{krige.mcgf()}} and \code{\link[=krige.mcgf_rs]{krige.mcgf_rs()}} for more details.
}
