% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige_new.R
\name{krige_new}
\alias{krige_new}
\title{Generic function for computing kriging forecasts for new locations}
\usage{
krige_new(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
Kriging results of \code{x}.
}
\description{
Generic function for computing kriging forecasts for new locations
}
\details{
Refer to \code{\link[=krige_new.mcgf]{krige_new.mcgf()}} and \code{\link[=krige_new.mcgf_rs]{krige_new.mcgf_rs()}} for more details.
}
