% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_joint.R
\name{mat_inv}
\alias{mat_inv}
\title{Find inverse of a symmetric positive definite matrix}
\usage{
mat_inv(x)
}
\arguments{
\item{x}{A symmetric and positive definite matrix.}
}
\value{
Inverse of x.
}
\description{
Find inverse of a symmetric positive definite matrix
}
\keyword{internal}
