% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Generic function for displaying fitted models for \code{mcgf} objects}
\usage{
model(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
Details of the fitted models.
}
\description{
Generic function for displaying fitted models for \code{mcgf} objects
}
\details{
Refer to \code{\link[=model.mcgf]{model.mcgf()}} and \code{\link[=model.mcgf_rs]{model.mcgf_rs()}} for more details.
}
