% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{obj_mle}
\alias{obj_mle}
\title{Compute the objective for mle method}
\usage{
obj_mle(par, cor_fn, x, lag, par_fixed)
}
\arguments{
\item{par}{Parameters of \code{cor_fn}.}

\item{cor_fn}{Correlation function}

\item{x}{An \code{mcgf} or \code{mcgf_rs} object}

\item{lag}{Time lag.}

\item{par_fixed}{Fixed parameters of \code{cor_fn}.}
}
\value{
A numeric value proportional to the negative approximate conditional
log-likelihood, to be minimized during parameter estimation.
}
\description{
Compute the objective for mle method
}
\keyword{internal}
