% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{obj_wls}
\alias{obj_wls}
\title{Compute the objective for wls method}
\usage{
obj_wls(par, cor_fn, cor_emp, par_fixed)
}
\arguments{
\item{par}{Parameters of \code{cor_fn}.}

\item{cor_fn}{Correlation function.}

\item{cor_emp}{Empirical correlations.}

\item{par_fixed}{Fixed parameters of \code{cor_fn}.}
}
\value{
The objective of weighted least squares.
}
\description{
Compute the objective for wls method
}
\keyword{internal}
