% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{optim_wls}
\alias{optim_wls}
\title{Optimization for wls method}
\usage{
optim_wls(par_init, optim_fn, cor_fn, cor_emp, par_fixed, lower, upper, ...)
}
\arguments{
\item{par_init}{Initial values for parameters to be optimized.}

\item{optim_fn}{Optimization function.}

\item{cor_fn}{Correlation function.}

\item{cor_emp}{Empirical correlations.}

\item{par_fixed}{Fixed parameters of \code{cor_fn}.}

\item{lower}{Lower bound.}

\item{upper}{Upper bound.}

\item{...}{Additional arguments passed to \code{optim_fn}.}
}
\value{
A list outputted from optimization functions of \code{optim_fn}.
}
\description{
Optimization for wls method
}
\keyword{internal}
