% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{rdists}
\alias{rdists}
\title{Generate random distance matrices}
\usage{
rdists(N, names, scale = 100)
}
\arguments{
\item{N}{Number of locations.}

\item{names}{Names of locations.}

\item{scale}{Scale of the distance matrices. Default is 100.}
}
\value{
List of signed distances.
}
\description{
Generate random distance matrices
}
\details{
This function generates random distance matrices using \code{rnorm}. \code{scale}
controls the scale of the distance matrices.
}
\examples{
set.seed(123)
rdists(3)
rdists(3, scale = 1)
rdists(3, names = LETTERS[1:3])
}
