% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds.R
\name{sds}
\alias{sds}
\title{Generic function for standard deviations for each column}
\usage{
sds(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
A vector of standard deviations for \code{mcgf} objects, or that plus
a list of regime-switching standard deviations for \code{mcgf_rs} objects.
}
\description{
Generic function for standard deviations for each column
}
\details{
Refer to \code{\link[=sds.mcgf]{sds.mcgf()}} and \code{\link[=sds.mcgf_rs]{sds.mcgf_rs()}} for more details.
}
