% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nugget.R
\name{set_nugget}
\alias{set_nugget}
\title{Adjust for nugget effect for correlations}
\usage{
set_nugget(x, nugget, set)
}
\arguments{
\item{x}{A correlation matrix or 3-d array.}

\item{nugget}{A scalar nugget effect.}

\item{set}{A correlation matrix or 3-d array of the same dimension as \code{x}.}
}
\value{
Correlations of the same dimension as \code{x}.
}
\description{
Adjust for nugget effect for correlations
}
\details{
The nugget effect is set to the correlations by first multiplying every
entry by \eqn{(1-\text{nugget})}. Then the diagonals (or the diagonals of
each matrix slice) of \code{x} are set to the corresponding diagonals of \code{set}.
}
\keyword{internal}
