% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_ar.R
\name{to_ar}
\alias{to_ar}
\title{Convert to array}
\usage{
to_ar(h, lag_max, u = TRUE)
}
\arguments{
\item{h}{Distance matrix.}

\item{lag_max}{Maximum time lag.}

\item{u}{Logical; TRUE if u_ar needs to be outputted.}
}
\value{
A list of arrays of h and u.
}
\description{
Convert to array
}
\keyword{internal}
