% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgf.R
\name{validate_mcgf}
\alias{validate_mcgf}
\title{Validate an mcgf object}
\usage{
validate_mcgf(x)
}
\arguments{
\item{x}{An mcgf object.}
}
\value{
An S3 object of class \code{mcgf}.
}
\description{
Validate an mcgf object
}
\details{
It validates an \code{mcgf} object by checking if \code{dists} contains valid
distance matrices/arrays.
}
\keyword{internal}
