% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgf_rs.R
\name{validate_mcgf_rs}
\alias{validate_mcgf_rs}
\title{Validate an mcgf_rs object}
\usage{
validate_mcgf_rs(x)
}
\arguments{
\item{x}{An mcgf_rs object}
}
\value{
An S3 object of class \code{mcgf_rs}.
}
\description{
Validate an mcgf_rs object
}
\details{
It validates an \code{mcgf_rs} object by checking if \code{label} is of the matching
length.
}
\keyword{internal}
