% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfs.R
\name{acfs.mcgf}
\alias{acfs.mcgf}
\alias{`acfs<-`}
\alias{acfs<-}
\title{Calculating mean autocorrelations for an \code{mcgf} object}
\usage{
\method{acfs}{mcgf}(x, lag_max, ...)

acfs(x) <- value
}
\arguments{
\item{x}{An \code{mcgf} object.}

\item{lag_max}{Maximum lag at which to calculate the acf.}

\item{...}{Additional parameters or attributes.}

\item{value}{A Vector of mean of autocorrelations for time lags starting
from 0.}
}
\value{
Mean autocorrelations.attribute \code{acfs}.
}
\description{
Calculating mean autocorrelations for an \code{mcgf} object
}
\details{
It computes mean autocorrelations for each time lag across locations. Use
\code{\link[=add_acfs]{add_acfs()}} to add \code{acfs} to \code{x}.
}
\examples{
wind_sq <- sqrt(wind[, -1])
time <- wind[, 1]
wind_mcgf <- mcgf(data = wind_sq, locations = wind_loc, time = time)
acfs(x = wind_mcgf, lag_max = 3)
}
\seealso{
Other {functions related to the class}: 
\code{\link{acfs}()},
\code{\link{add_acfs.mcgf}()},
\code{\link{add_acfs}()},
\code{\link{add_ccfs.mcgf}()},
\code{\link{add_ccfs}()},
\code{\link{ccfs.mcgf}()},
\code{\link{ccfs}()},
\code{\link{dists.mcgf}()},
\code{\link{dists}()},
\code{\link{fit_base}()},
\code{\link{mcgf}()},
\code{\link{print.mcgf}()}
}
\concept{{functions related to the class}}
