% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfs.R
\name{acfs}
\alias{acfs}
\title{Generic function for calculating autocorrelation}
\usage{
acfs(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
Mean of autocorrelations for each time lag.
}
\description{
Generic function for calculating autocorrelation
}
\seealso{
Other {functions related to the class}: 
\code{\link{acfs.mcgf}()},
\code{\link{add_acfs.mcgf}()},
\code{\link{add_acfs}()},
\code{\link{add_ccfs.mcgf}()},
\code{\link{add_ccfs}()},
\code{\link{ccfs.mcgf}()},
\code{\link{ccfs}()},
\code{\link{dists.mcgf}()},
\code{\link{dists}()},
\code{\link{fit_base}()},
\code{\link{mcgf}()},
\code{\link{print.mcgf}()},
\code{\link{sds}()}
}
\concept{{functions related to the class}}
