% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfs.R
\name{acfs}
\alias{acfs}
\title{Generic function for calculating autocorrelation}
\usage{
acfs(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\description{
Generic function for calculating autocorrelation
}
\details{
Refer to \code{\link[=acfs.mcgf]{acfs.mcgf()}} and \code{\link[=acfs.mcgf_rs]{acfs.mcgf_rs()}} for more details.
}
