% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfs.R
\name{add_acfs.mcgf}
\alias{add_acfs.mcgf}
\title{Calculating and adding mean autocorrelations for an \code{mcgf} object}
\usage{
\method{add_acfs}{mcgf}(x, lag_max, ...)
}
\arguments{
\item{x}{An \code{mcgf} object.}

\item{lag_max}{Maximum lag at which to calculate the acf.}

\item{...}{Additional parameters or attributes.}
}
\value{
An \code{mcgf} object with newly added mean autocorrelations under
attribute \code{acfs}.
}
\description{
Calculating and adding mean autocorrelations for an \code{mcgf} object
}
\details{
It computes mean autocorrelations for each time lag across locations, and add
them to \code{x}.
}
\examples{
wind_sq <- sqrt(wind[, -1])
time <- wind[, 1]
wind_mcgf <- mcgf(data = wind_sq, locations = wind_loc, time = time)
wind_mcgf <- add_acfs(x = wind_mcgf, lag_max = 3)
print(wind_mcgf, "acfs")
}
\seealso{
Other {functions related to the class}: 
\code{\link{acfs.mcgf}()},
\code{\link{acfs}()},
\code{\link{add_acfs}()},
\code{\link{add_ccfs.mcgf}()},
\code{\link{add_ccfs}()},
\code{\link{ccfs.mcgf}()},
\code{\link{ccfs}()},
\code{\link{dists.mcgf}()},
\code{\link{dists}()},
\code{\link{fit_base}()},
\code{\link{mcgf}()},
\code{\link{print.mcgf}()},
\code{\link{sds}()}
}
\concept{{functions related to the class}}
