% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfs.R
\name{add_acfs}
\alias{add_acfs}
\title{Generic function for calculating and adding autocorrelation}
\usage{
add_acfs(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
\code{x} with mean of autocorrelations for each time lag.
}
\description{
Generic function for calculating and adding autocorrelation
}
\seealso{
Other {functions related to the auto- and cross-correlations}: 
\code{\link{acfs.mcgf}()},
\code{\link{acfs}()},
\code{\link{add_acfs.mcgf}()},
\code{\link{add_ccfs.mcgf}()},
\code{\link{add_ccfs}()},
\code{\link{ccfs.mcgf}()},
\code{\link{ccfs}()}
}
\concept{{functions related to the auto- and cross-correlations}}
