% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfs.R
\name{add_acfs}
\alias{add_acfs}
\title{Generic function for calculating autocorrelation}
\usage{
add_acfs(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
Mean of auto-correlations.
}
\description{
Generic function for calculating autocorrelation
}
\seealso{
Other {functions related to the class}: 
\code{\link{add_acfs.mcgf}()},
\code{\link{dists.mcgf}()},
\code{\link{dists}()},
\code{\link{mcgf}()},
\code{\link{print.mcgf}()}
}
\concept{{functions related to the class}}
