% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_base.R
\name{add_base.mcgf}
\alias{add_base.mcgf}
\alias{base<-}
\title{Add base model outputted from \code{fit_base} to a mcgf object.}
\usage{
\method{add_base}{mcgf}(x, fit_base, ...)

base(x) <- value
}
\arguments{
\item{x}{An mcgf object.}

\item{fit_base}{Output from the \code{fit_base} function.}

\item{...}{Additional arguments. Not in use.}

\item{value}{A list containing the base model as well as its parameters. It
must contains \code{model}, \code{par_base}, \code{cov_base}, \code{lag_base}, and \code{horizon}.}
}
\value{
\code{x} with newly added attributes of the base model.
}
\description{
Add base model outputted from \code{fit_base} to a mcgf object.
}
\seealso{
Other {functions related to model fitting}: 
\code{\link{add_base}()},
\code{\link{fit_base.mcgf}()},
\code{\link{fit_base}()}
}
\concept{{functions related to model fitting}}
