% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_base.R
\name{add_base.mcgf_rs}
\alias{add_base.mcgf_rs}
\title{Add base model outputted from \code{\link[=fit_base]{fit_base()}} to an \code{mcgf_rs} object.}
\usage{
\method{add_base}{mcgf_rs}(x, fit_base_ls, fit_s_ls, fit_t_ls, sep = FALSE, old = FALSE, ...)
}
\arguments{
\item{x}{An mcgf_rs` object.}

\item{fit_base_ls}{Output from the \code{\link[=fit_base]{fit_base()}} function.}

\item{fit_s_ls}{Pure spatial model outputted from the \code{\link[=fit_base]{fit_base()}} function.
Used only when \code{sep = TRUE}.}

\item{fit_t_ls}{Pure temporal model outputted from the \code{\link[=fit_base]{fit_base()}}
function. Used only when \code{sep = TRUE}.}

\item{sep}{Logical; TRUE if spatial and temporal models are fitted
separately.}

\item{old}{Logical; TRUE if the old base model needs to be kept. The lag and
horizon of the new model are assumed to be the same as that of the old model.}

\item{...}{Additional arguments. Not in use.}
}
\value{
\code{x} with newly added attributes of the base model.
}
\description{
Add base model outputted from \code{\link[=fit_base]{fit_base()}} to an \code{mcgf_rs} object.
}
\details{
After fitting the base model by \code{\link[=fit_base]{fit_base()}}, the results can be added to
\code{x} by \code{\link[=add_base]{add_base()}}. To supply the base model directly, use \code{\link{base<-}} to
add the base model; the value must contain the same output as
\code{\link[=add_base.mcgf]{add_base.mcgf()}} or \code{\link[=add_base.mcgf_rs]{add_base.mcgf_rs()}}.
}
\seealso{
Other {functions related to model fitting}: 
\code{\link{add_base.mcgf}()},
\code{\link{add_lagr.mcgf_rs}()},
\code{\link{add_lagr.mcgf}()},
\code{\link{add_lagr}()},
\code{\link{fit_base.mcgf_rs}()},
\code{\link{fit_base.mcgf}()},
\code{\link{fit_base}()},
\code{\link{fit_lagr.mcgf_rs}()},
\code{\link{fit_lagr.mcgf}()},
\code{\link{fit_lagr}()},
\code{\link{model.mcgf_rs}()},
\code{\link{model.mcgf}()},
\code{\link{model}()}
}
\concept{{functions related to model fitting}}
