% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_base.R
\name{add_base}
\alias{add_base}
\title{Generic function for adding a base model}
\usage{
add_base(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
Base model and its parameter
}
\description{
Generic function for adding a base model
}
\seealso{
Other {functions related to model fitting}: 
\code{\link{add_base.mcgf}()},
\code{\link{add_lagr.mcgf}()},
\code{\link{add_lagr}()},
\code{\link{fit_base.mcgf}()},
\code{\link{fit_base}()},
\code{\link{fit_lagr.mcgf}()},
\code{\link{fit_lagr}()},
\code{\link{model.mcgf}()},
\code{\link{model}()}
}
\concept{{functions related to model fitting}}
