% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lagr.R
\name{add_lagr.mcgf}
\alias{add_lagr.mcgf}
\alias{lagr<-}
\title{Add lagr model outputted from \code{\link[=fit_lagr]{fit_lagr()}} to a \code{mcgf} object.}
\usage{
\method{add_lagr}{mcgf}(x, fit_lagr, ...)

lagr(x) <- value
}
\arguments{
\item{x}{An \code{mcgf} object.}

\item{fit_lagr}{Output from the \code{\link[=fit_lagr]{fit_lagr()}} function.}

\item{...}{Additional arguments. Not in use.}

\item{value}{A list containing the lagr model as well as its parameters. It
must contains \code{model}, \code{par_lagr}, and \code{cor_lagr}.}
}
\value{
\code{x} with newly added attributes of the Lagrangian model.
}
\description{
Add lagr model outputted from \code{\link[=fit_lagr]{fit_lagr()}} to a \code{mcgf} object.
}
\seealso{
Other {functions related to model fitting}: 
\code{\link{add_base.mcgf_rs}()},
\code{\link{add_base.mcgf}()},
\code{\link{add_lagr.mcgf_rs}()},
\code{\link{add_lagr}()},
\code{\link{fit_base.mcgf_rs}()},
\code{\link{fit_base.mcgf}()},
\code{\link{fit_base}()},
\code{\link{fit_lagr.mcgf_rs}()},
\code{\link{fit_lagr.mcgf}()},
\code{\link{fit_lagr}()},
\code{\link{model.mcgf_rs}()},
\code{\link{model.mcgf}()},
\code{\link{model}()}
}
\concept{{functions related to model fitting}}
