% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lagr.R
\name{add_lagr.mcgf_rs}
\alias{add_lagr.mcgf_rs}
\title{Add lagr model outputted from \code{\link[=fit_lagr]{fit_lagr()}} to a \code{mcgf_rs} object.}
\usage{
\method{add_lagr}{mcgf_rs}(x, fit_lagr_ls, ...)
}
\arguments{
\item{x}{An \code{mcgf_rs} object.}

\item{fit_lagr_ls}{Output from the \code{\link[=fit_lagr]{fit_lagr()}} function.}

\item{...}{Additional arguments. Not in use.}
}
\value{
\code{x} with newly added attributes of the Lagrangian model.
}
\description{
Add lagr model outputted from \code{\link[=fit_lagr]{fit_lagr()}} to a \code{mcgf_rs} object.
}
\details{
After fitting the Lagrangian model by \code{\link[=fit_lagr]{fit_lagr()}}, the results can be
added to \code{x} by \code{\link[=add_base]{add_base()}}. To supply the Lagrangian model directly,
use \code{\link{lagr<-}} to add the Lagrangian model; the value must contain the same
output as \code{\link[=add_lagr.mcgf]{add_lagr.mcgf()}} or \code{\link[=add_lagr.mcgf_rs]{add_lagr.mcgf_rs()}}.
}
\seealso{
Other {functions related to model fitting}: 
\code{\link{add_base.mcgf_rs}()},
\code{\link{add_base.mcgf}()},
\code{\link{add_lagr.mcgf}()},
\code{\link{add_lagr}()},
\code{\link{fit_base.mcgf_rs}()},
\code{\link{fit_base.mcgf}()},
\code{\link{fit_base}()},
\code{\link{fit_lagr.mcgf_rs}()},
\code{\link{fit_lagr.mcgf}()},
\code{\link{fit_lagr}()},
\code{\link{model.mcgf_rs}()},
\code{\link{model.mcgf}()},
\code{\link{model}()}
}
\concept{{functions related to model fitting}}
