% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lagr.R
\name{add_lagr}
\alias{add_lagr}
\title{Generic function for adding a Lagrangian model}
\usage{
add_lagr(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
lagr model and its parameter
}
\description{
Generic function for adding a Lagrangian model
}
\seealso{
Other {functions related to model fitting}: 
\code{\link{add_base.mcgf_rs}()},
\code{\link{add_base.mcgf}()},
\code{\link{add_lagr.mcgf_rs}()},
\code{\link{add_lagr.mcgf}()},
\code{\link{fit_base.mcgf_rs}()},
\code{\link{fit_base.mcgf}()},
\code{\link{fit_base}()},
\code{\link{fit_lagr.mcgf_rs}()},
\code{\link{fit_lagr.mcgf}()},
\code{\link{fit_lagr}()},
\code{\link{model.mcgf_rs}()},
\code{\link{model.mcgf}()},
\code{\link{model}()}
}
\concept{{functions related to model fitting}}
