% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nugget.R
\name{add_nugget}
\alias{add_nugget}
\title{Adjust for nugget effect for correlations}
\usage{
add_nugget(x, nugget)
}
\arguments{
\item{x}{A correlation matrix or 3-d array.}

\item{nugget}{A scalar nugget effect.}
}
\value{
Correlations of the same dimension as \code{x}.
}
\description{
Adjust for nugget effect for correlations
}
\details{
The nugget effect is added to the correlations by first multiplying every
entry by \eqn{(1-\text{nugget})}. Then the diagonals (or the diagonals of
each matrix slice) of \code{x} are set to 1 based on the assumption that the
correlation is 1 when \code{x} is 0.
}
\keyword{internal}
