% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfs.R
\name{ccf_rs}
\alias{ccf_rs}
\title{Calculate regime-switching cross-correlation}
\usage{
ccf_rs(x, y, label, lag_max)
}
\arguments{
\item{x, y}{A univariate numeric time series.}

\item{label}{A vector of regime labels.}

\item{lag_max}{Maximum lag at which to calculate the ccf.}
}
\value{
Cross-correlations for each group in \code{label}.
}
\description{
Calculate regime-switching cross-correlation
}
