% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfs.R
\name{ccfs.mcgf}
\alias{ccfs.mcgf}
\alias{ccfs.mcgf_rs}
\alias{ccfs<-}
\alias{add_ccfs}
\title{Extract, calculate, or assign cross-correlations for an \code{mcgf} or \code{mcgf_rs}
object}
\usage{
\method{ccfs}{mcgf}(x, lag_max, ncores = 1, replace = FALSE, ...)

\method{ccfs}{mcgf_rs}(x, lag_max, ncores = 1, replace = FALSE, ...)

ccfs(x) <- value

add_ccfs(x, lag_max, ncores = 1, ...)
}
\arguments{
\item{x}{An \code{mcgf} or \code{mcgf_rs} object.}

\item{lag_max}{Maximum lag at which to calculate the ccfs.}

\item{ncores}{Number of cpu cores used for computing. The \code{doParallel}
package is required when \code{ncores} > 1.}

\item{replace}{Logical; if TRUE, \code{acfs} are recalculated.}

\item{...}{Additional parameters or attributes. Not in use.}

\item{value}{Cross-correlations.}
}
\value{
\code{\link[=ccfs]{ccfs()}} returns (regime-switching) cross-correlations.
\code{\link[=add_ccfs]{add_ccfs()}} returns the same object with additional attributes of
(regime-switching) cross-correlations and (regime-switching) empirical
standard deviations.
}
\description{
Extract, calculate, or assign cross-correlations for an \code{mcgf} or \code{mcgf_rs}
object
}
\details{
For \code{mcgf} objects, \code{\link[=ccfs]{ccfs()}} computes cross-correlations for each time
lag. The output is an array of matrices where each matrix corresponds to the
cross-correlation for a time lag.

For \code{mcgf_rs} objects, \code{\link[=ccfs]{ccfs()}} computes regime-switching
cross-correlations for each time lag. The output is a list of array of
matrices where each array in the list corresponds to the ccfs for a regime.

\code{\link{ccfs<-}} assigns \code{ccfs} to \code{x}.

\code{\link[=add_ccfs]{add_ccfs()}} adds \code{ccfs} and \code{sds} to \code{x}.
}
\examples{
wind_sq <- sqrt(wind[, -1])
time <- wind[, 1]
wind_mcgf <- mcgf(data = wind_sq, locations = wind_loc, time = time)
ccfs(x = wind_mcgf, lag_max = 3)
# ccfs(x = wind_mcgf, lag_max = 3, ncores = 10)

wind_sq <- sqrt(wind[, -1])
time <- wind[, 1]
wind_mcgf <- mcgf_rs(
    data = wind_sq, locations = wind_loc, time = time,
    label = c(rep(1, 3574), rep(2, 3000))
)
ccfs(x = wind_mcgf, lag_max = 3)
# ccfs(x = wind_mcgf, lag_max = 3, ncores = 10)
wind_sq <- sqrt(wind[, -1])
time <- wind[, 1]
wind_mcgf <- mcgf(data = wind_sq, locations = wind_loc, time = time)
wind_mcgf <- add_ccfs(x = wind_mcgf, lag_max = 3)
print(wind_mcgf, "ccfs")
print(wind_mcgf, "sds")
}
\seealso{
Other {functions related to the auto- and cross-correlations}: 
\code{\link{acfs.mcgf}()}
}
\concept{{functions related to the auto- and cross-correlations}}
