% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfs.R, R/sds.R
\name{ccfs.mcgf}
\alias{ccfs.mcgf}
\alias{`ccfs<-`}
\alias{ccfs<-}
\alias{sds.mcgf}
\title{Calculating cross-correlations for an \code{mcgf} object}
\usage{
\method{ccfs}{mcgf}(x, lag_max, ...)

ccfs(x) <- value

\method{sds}{mcgf}(x, ...)
}
\arguments{
\item{x}{An \code{mcgf} object.}

\item{lag_max}{Maximum lag at which to calculate the acf.}

\item{...}{Additional parameters or attributes.}

\item{value}{Cross-correlations.}
}
\value{
Cross-correlations.
}
\description{
Calculating cross-correlations for an \code{mcgf} object
}
\details{
It computes cross-correlations for each time lag. The output is an array of
matrices where each matrix corresponds to the cross-correlation for a time
lag. Use \code{\link[=add_ccfs]{add_ccfs()}} to add \code{ccfs} and \code{sds} to \code{x}.
}
\examples{
wind_sq <- sqrt(wind[, -1])
time <- wind[, 1]
wind_mcgf <- mcgf(data = wind_sq, locations = wind_loc, time = time)
ccfs(x = wind_mcgf, lag_max = 3)

}
\seealso{
Other {functions related to the auto- and cross-correlations}: 
\code{\link{acfs.mcgf}()},
\code{\link{acfs}()},
\code{\link{add_acfs.mcgf}()},
\code{\link{add_acfs}()},
\code{\link{add_ccfs.mcgf}()},
\code{\link{add_ccfs}()},
\code{\link{ccfs}()}
}
\concept{{functions related to the auto- and cross-correlations}}
