% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccfs.R
\name{ccfs}
\alias{ccfs}
\title{Generic function for calculating cross-autocorrelation}
\usage{
ccfs(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
Cross-correlations.
}
\description{
Generic function for calculating cross-autocorrelation
}
\seealso{
Other {functions related to the auto- and cross-correlations}: 
\code{\link{acfs.mcgf}()},
\code{\link{acfs}()},
\code{\link{add_acfs.mcgf}()},
\code{\link{add_acfs}()},
\code{\link{add_ccfs.mcgf}()},
\code{\link{add_ccfs}()},
\code{\link{ccfs.mcgf}()}
}
\concept{{functions related to the auto- and cross-correlations}}
