% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccov.R
\name{ccov.mcgf}
\alias{ccov.mcgf}
\title{Covariance/correlation for joint distribution of an mcgf object}
\usage{
\method{ccov}{mcgf}(x, model = c("all", "base", "empirical"), cor = FALSE, ...)
}
\arguments{
\item{x}{An mcgf object.}

\item{model}{Which model to use. One of \code{all}, \code{base}, and \code{empirical}.}

\item{cor}{Logical; if TRUE, correlation is outputted.}

\item{...}{Additional arguments. Not in use.}
}
\value{
Joint correlation matrix.
}
\description{
Covariance/correlation for joint distribution of an mcgf object
}
