% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccov.R
\name{ccov.mcgf_rs}
\alias{ccov.mcgf_rs}
\title{Covariance/correlation for joint distribution of an \code{mcgf_rs}object}
\usage{
\method{ccov}{mcgf_rs}(x, model = c("all", "base", "empirical"), cor = FALSE, ...)
}
\arguments{
\item{x}{An \code{mcgf} object.}

\item{model}{Which model to use. One of \code{all}, \code{base}, and \code{empirical}.}

\item{cor}{Logical; if TRUE, correlation is returned}

\item{...}{Additional arguments. Not in use.}
}
\value{
A list of joint vovariance/correlation matrix.
}
\description{
Covariance/correlation for joint distribution of an \code{mcgf_rs}object
}
