% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dist}
\alias{check_dist}
\title{Check if valid distance}
\usage{
check_dist(x, name = "x", check_sym = TRUE)
}
\arguments{
\item{x}{Distance matrix or array.}

\item{name}{Name of \code{x} for displaying errors.}

\item{check_sym}{Logical; if TRUE each matrix (slice) must be symmetric.}
}
\value{
NULL.
}
\description{
Check if valid distance
}
\details{
Check if \code{x} is a valid distance vector, matrix or array. It errors if any
elements in \code{x} is negative, or if \code{x} is not a symmetric matrix or an
array of symmetric matrices.
}
\keyword{internal}
