% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_dist}
\alias{check_dist}
\title{Check if valid distance}
\usage{
check_dist(x)
}
\arguments{
\item{x}{Distance matrix or array.}
}
\value{
\code{x}.
}
\description{
Check if valid distance
}
\details{
Check if input \code{x} is a valid distance vector, matrix or array. It errors if
any elements in \code{x} is negative, or if \code{x} is not a symmetric matrix or an
array of symmetric matrices.
}
\keyword{internal}
