% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_length}
\alias{check_length}
\title{Check if valid input length}
\usage{
check_length(x, length, name)
}
\arguments{
\item{x}{Scaler or vector}

\item{length}{Length of \code{x}.}

\item{name}{Name of \code{x} for displaying errors.}
}
\value{
\code{x}.
}
\description{
Check if valid input length
}
\details{
Check if \code{x} has approprate length. If length of \code{x} is 1 then \code{x} is
replicated to match \code{length}. If length of \code{x} is neither 1 or \code{length} then
an error is signaled.
}
\keyword{internal}
