% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_length_ls}
\alias{check_length_ls}
\title{Check if valid input length}
\usage{
check_length_ls(x_ls, length, name)
}
\arguments{
\item{x_ls}{List of scaler or vector}

\item{length}{List of length of \code{x_ls}.}

\item{name}{Name of \code{x} for displaying errors.}
}
\value{
\code{x_ls}.
}
\description{
Check if valid input length
}
\details{
Check if elements in \code{x_ls} have approprate length. If length of any elements
in \code{x_ls} is 1 then they are replicated to match \code{length}. If length of any
elements is neither 1 or \code{length} then an error is signaled.
}
\keyword{internal}
