% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2cov.R
\name{cor2cov}
\alias{cor2cov}
\title{Convert correlation to covariance}
\usage{
cor2cov(V, sd)
}
\arguments{
\item{V}{A correlation matrix, usually positive semi-definite.}

\item{sd}{A vector of standard deviations.}
}
\value{
A correlation matrix.
}
\description{
Convert correlation to covariance
}
\examples{
V <- matrix(c(1, 0.5, 0.5, 1), ncol = 2)
sd <- 1:2
cor2cov(V, sd)
}
