% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_cauchy.R
\name{cor_cauchy}
\alias{cor_cauchy}
\title{Calculate Cauchy correlation}
\usage{
cor_cauchy(nugget = 0, a, alpha, x, is.dist = F)
}
\arguments{
\item{nugget}{The nugget effect in \eqn{\in[0, 1]}.}

\item{a}{Smooth parameter, a>0}

\item{alpha}{Scale parameter, \eqn{\alpha\in(0, 1]}.}

\item{x}{A vector, matrix, or array.}

\item{is.dist}{Logical; if TRUE, x is a distance matrix or an array of distance matrices.}
}
\value{
sf
}
\description{
Calculate Cauchy correlation
}
