% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_lagr_tri.R
\name{cor_lagr_tri}
\alias{cor_lagr_tri}
\title{Calculate Lagrangian correlation of the triangular form}
\usage{
cor_lagr_tri(v1, v2, h1, h2, u)
}
\arguments{
\item{v1}{Prevailing wind, u-component.}

\item{v2}{Prevailing wind, v-component.}

\item{h1}{Horizontal distance matrix or array.}

\item{h2}{Vertical distance matrix or array, same dimension as \code{h1}.}

\item{u}{Time lag, same dimension as \code{h1}.}
}
\value{
Correlations of the same dimension as \code{h1}.
}
\description{
Calculate Lagrangian correlation of the triangular form
}
\details{
The Lagrangian correlation function of the triangular form with parameters
\eqn{\boldsymbol v = (v_1, v_2)^\top} has the form
\deqn{C(\mathbf{h}, u)=\left(1-\dfrac{1}{2\|\boldsymbol v\|}
\left|\dfrac{\mathbf{h}^\top\boldsymbol v}{\|\boldsymbol v\|}-
u\|\boldsymbol v\|\right|\right)_+,}
where \eqn{x_+=\max(x,0).}
}
\examples{
h1 <- matrix(c(0, 5, -5, 0), nrow = 2)
h2 <- matrix(c(0, 8, -8, 0), nrow = 2)
u <- matrix(0.1, nrow = 2, ncol = 2)
cor_lagr_tri(v1 = 5, v = 10, h1 = h1, h2 = h2, u = u)

h1 <- array(c(0, 5, -5, 0), dim = c(2, 2, 3))
h2 <- array(c(0, 8, -8, 0), dim = c(2, 2, 3))
u <- array(rep(c(0.1, 0.2, 0.3), each = 4), dim = c(2, 2, 3))
cor_lagr_tri(v1 = 5, v = 10, h1 = h1, h2 = h2, u = u)

}
\seealso{
\link{cor_exp}, \link{cor_cauchy}, \link{cor_fs}
}
