% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_sep.R
\name{cor_sep}
\alias{cor_sep}
\title{Calculate correlation for separable model}
\usage{
cor_sep(
  par_s,
  par_t,
  h,
  u,
  spatial = c("exp", "cauchy"),
  temporal = c("exp", "cauchy")
)
}
\arguments{
\item{par_s}{Parameters for the pure spatial model. Nugget effect supported.}

\item{par_t}{Parameters for the pure temporal model.}

\item{h}{Euclidean distance matrix or array.}

\item{u}{Time lag, same dimension as \code{h}.}

\item{spatial}{Pure spatial model, \code{exp} or \code{cauchy} for now.}

\item{temporal}{Pure temporal model, \code{exp} or \code{cauchy} for now.}
}
\value{
Correlations of the same dimension as u
}
\description{
Calculate correlation for separable model
}
\details{
The separable model is the product of a pure spatial model and pure temporal
model. Now only exponential and Cauchy correlation models are available.
}
\examples{
h <- matrix(c(0, 5, 5, 0), nrow = 2)
par_s <- list(nugget = 0.5, c = 0.01, gamma = 0.5)
u <- matrix(0, nrow = 2, ncol = 2)
par_t <- list(a = 1, alpha = 0.5)
cor_sep(par_s = par_s, par_t = par_t, h = h, u = u, spatial = "exp",
        temporal = "cauchy")

h <- array(c(0, 5, 5, 0), dim = c(2, 2, 3))
par_s <- list(nugget = 0.5, c = 0.01, gamma = 0.5)
u <- array(rep(0:2, each = 4), dim = c(2, 2, 3))
par_t <- list(a = 1, alpha = 0.5)
cor_sep(par_s = par_s, par_t = par_t, h = h, u = u, spatial = "exp",
        temporal = "cauchy")

}
\references{
Gneiting, T. (2002). Nonseparable, Stationary Covariance Functions for
Space–Time Data, Journal of the American Statistical Association, 97:458,
590-600.
}
\seealso{
\link{cor_exp}, \link{cor_cauchy}, \link{cor_fs}, \link{cor_lagr_tri}
}
