% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_joint.R
\name{cov_joint}
\alias{cov_joint}
\title{Covariance for joint distribution}
\usage{
cov_joint(cov)
}
\arguments{
\item{cov}{Array of covariance matrices.}
}
\value{
A covariance matrix for the joint distribution of the current value
and the past value for a Markov chain Gaussian field.
}
\description{
Covariance for joint distribution
}
\details{
The covariance matrix of the joint distribution has the block Toeplitz
structure. Input \code{cov} is assumed to be an array of covariance matrices
where the ith matrix slice correspond to the i-1 time lag. For example,
\code{cov[, , 1]} is the covariance matrix for time lag 0. All matrices in \code{cov}
are used to construct the joint covariance matrix.
}
\examples{
1
}
