% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_sep.R
\name{.cor_sep}
\alias{.cor_sep}
\title{Calculate correlation for separable model}
\usage{
.cor_sep(par_s, par_t, spatial, temporal)
}
\arguments{
\item{par_s}{Parameters for the pure spatial model. Nugget effect supported.}

\item{par_t}{Parameters for the pure temporal model.}

\item{spatial}{Pure spatial model, \code{exp} or \code{cauchy} for now.}

\item{temporal}{Pure temporal model, \code{exp} or \code{cauchy} for now.}
}
\value{
Correlations for separable model.
}
\description{
Calculate correlation for separable model
}
\details{
The separable model is the product of a pure spatial model and pure temporal
model. Now only exponential and Cauchy correlation models are available.
}
\keyword{internal}
