% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_stat.R
\name{.corr_stat}
\alias{.corr_stat}
\title{Calculate general stationary correlation.}
\usage{
.corr_stat(
  base,
  dist,
  lag_max,
  par,
  lambda,
  asymm_model = c("none", "lagr_tri")
)
}
\arguments{
\item{base}{Matrix or array of correlations.}

\item{dist}{List of distance matrices or arrays.}

\item{lag_max}{Maximum lag considered for the Lagrangian term.}

\item{par}{List of parameters for the Lagrangian term.}

\item{asymm_model}{}
}
\description{
Calculate general stationary correlation.
}
