% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dists.R
\name{.find_dists}
\alias{.find_dists}
\title{Calculate (signed) distances between coordinates}
\usage{
.find_dists(grid, names = NULL, longlat = TRUE)
}
\arguments{
\item{grid}{A matrix of 2D points, first column x/longitude, second column
y/latitude.}

\item{names}{Names of locations.}

\item{longlat}{Logical, if TURE Great Circle (WGS84 ellipsoid) distance;
if FALSE, Euclidean distance.}
}
\value{
List of signed distances.
}
\description{
Calculate (signed) distances between coordinates
}
\keyword{internal}
