% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_base.R
\name{fit_base.mcgf}
\alias{fit_base.mcgf}
\title{Parameter estimation for symmetric correlation functions.}
\usage{
\method{fit_base}{mcgf}(
  x,
  lag,
  horizon = 1,
  model = c("spatial", "temporal", "sep", "fs"),
  method = c("wls", "mle"),
  optim_fn = c("nlminb", "optim", "other"),
  par_fixed,
  par_init,
  lower,
  upper,
  other_optim_fn,
  ...
)
}
\arguments{
\item{x}{An \code{mcgf} object containing attributes \code{dists}, \code{acfs}, \code{ccfs}, and
\code{sds}.}

\item{lag}{Integer time lag.}

\item{horizon}{Integer forecast horizon.}

\item{model}{Base model, one of \code{spatial}, \code{temporal}, \code{sep}, \code{fs}. Only
\code{sep} and \code{fs} are supported when \code{method = mle}}

\item{method}{Parameter estimation methods, weighted least square (\code{wls}) or
maximum likelihood estimation (\code{mle}).}

\item{optim_fn}{Optimization functions, one of \code{nlminb}, \code{optim}, \code{other}.
When \code{optim_fn = other}, supply \code{other_optim_fn}.}

\item{par_fixed}{Fixed parameters.}

\item{par_init}{Initial values for parameters to be optimized.}

\item{lower}{Optional; lower bound of parameters.}

\item{upper}{Optional: upper bound of parameters.}

\item{other_optim_fn}{Optional, other optimization functions. The first two
arguments must be initial values for the parameters and a function to be
minimized respectively (same as that of \code{optim} and \code{nlminb}).}

\item{...}{Additional arguments passed to \code{optim_fn}.}
}
\value{
A list outputted from optimization functions of \code{optim_fn}.
}
\description{
Parameter estimation for symmetric correlation functions.
}
\details{
Additional details...
}
