% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_base.R
\name{fit_base}
\alias{fit_base}
\title{Fit correlation models}
\usage{
fit_base(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
A vector of estimated parameters
}
\description{
Fit correlation models
}
\seealso{
Other {functions related to the class}: 
\code{\link{acfs.mcgf}()},
\code{\link{acfs}()},
\code{\link{add_acfs.mcgf}()},
\code{\link{add_acfs}()},
\code{\link{add_ccfs.mcgf}()},
\code{\link{add_ccfs}()},
\code{\link{ccfs.mcgf}()},
\code{\link{ccfs}()},
\code{\link{dists.mcgf}()},
\code{\link{dists}()},
\code{\link{mcgf}()},
\code{\link{print.mcgf}()},
\code{\link{sds}()}
}
\concept{{functions related to the class}}
