% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lagr.R
\name{fit_lagr.mcgf}
\alias{fit_lagr.mcgf}
\title{Parameter estimation for Lagrangian correlation functions.}
\usage{
\method{fit_lagr}{mcgf}(
  x,
  model = c("lagr_tri", "lagr_askey"),
  method = c("wls", "mle"),
  optim_fn = c("nlminb", "optim", "other"),
  par_fixed,
  par_init,
  lower,
  upper,
  other_optim_fn,
  ...
)
}
\arguments{
\item{x}{An \code{mcgf} object containing attributes \code{dists}, \code{acfs}, \code{ccfs}, and
\code{sds}. \code{x} must have been passed to \code{add_base()} or \verb{base<-}}

\item{model}{Base model, one of \code{lagr_tri}, \code{lagr_askey}.}

\item{method}{Parameter estimation methods, weighted least square (\code{wls}) or
maximum likelihood estimation (\code{mle}).}

\item{optim_fn}{Optimization functions, one of \code{nlminb}, \code{optim}, \code{other}.
When \code{optim_fn = other}, supply \code{other_optim_fn}.}

\item{par_fixed}{Fixed parameters.}

\item{par_init}{Initial values for parameters to be optimized.}

\item{lower}{Optional; lower bound of parameters.}

\item{upper}{Optional: upper bound of parameters.}

\item{other_optim_fn}{Optional, other optimization functions. The first two
arguments must be initial values for the parameters and a function to be
minimized respectively (same as that of \code{optim} and \code{nlminb}).}

\item{...}{Additional arguments passed to \code{optim_fn}.}
}
\value{
A list containing outputs from optimization functions of \code{optim_fn}.
}
\description{
Parameter estimation for Lagrangian correlation functions.
}
\details{
This function fits the Lagrangian models using weighted least squares and
maximum likelihood estimation. The base model must be fitted first using
\code{add_base()} or \verb{base<-}. Optimization functions such as \code{nlminb} and \code{optim}
are supported. Other functions are also supported by setting
\code{optim_fn = "other"} and supplying \code{other_optim_fn}. \code{lower} and \code{upper} are
lower and upper bounds of parameters in \code{par_init} and default bounds are
used if they are not specified.
}
\seealso{
Other {functions related to model fitting}: 
\code{\link{add_base.mcgf}()},
\code{\link{add_base}()},
\code{\link{add_lagr.mcgf}()},
\code{\link{add_lagr}()},
\code{\link{fit_base.mcgf}()},
\code{\link{fit_base}()},
\code{\link{fit_lagr}()},
\code{\link{model.mcgf}()},
\code{\link{model}()}
}
\concept{{functions related to model fitting}}
