% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgf_rs.R
\name{is.mcgf_rs}
\alias{is.mcgf_rs}
\alias{as.mcgf_rs}
\title{Check if an object is an \code{mcgf_rs} object..}
\usage{
is.mcgf_rs(x)

as.mcgf_rs(x, label)
}
\arguments{
\item{x}{An Object.}

\item{label}{A vector of regime labels. Its length must be the same as
the number rows in \code{data}.}
}
\value{
\code{is.mcgf_rs} returns a logical valud; TRUE if \code{x} is of the \code{mcgf_rs}
class. \code{as.mcgf_rs} coerces an \code{mcgf} object to an \code{mcgf_rs} object by adding
regime labels.
}
\description{
Check if an object is an \code{mcgf_rs} object..
}
