% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige.R
\name{krige.mcgf_rs}
\alias{krige.mcgf_rs}
\title{Obtain kriging forecasts for an \code{mcgf_rs} object.}
\usage{
\method{krige}{mcgf_rs}(
  x,
  newdata,
  newlabel,
  soft = FALSE,
  prob,
  model = c("all", "base", "empirical"),
  interval = FALSE,
  level = 0.95,
  ...
)
}
\arguments{
\item{x}{An \code{mcgf_rs} object.}

\item{newdata}{A data.frame with the same column names as \code{x}. If \code{newdata}
is missing the forecasts at the original data points are returned.}

\item{newlabel}{A vector of new regime labels.}

\item{soft}{Logical; if true, soft forecasts (and bounds) are produced.}

\item{prob}{Matrix with simplex rows. Number of columns must be the same as
unique labels in \code{x}.}

\item{model}{Which model to use. One of \code{all}, \code{base}, and \code{empirical}.}

\item{interval}{Logical; if TRUE, prediction intervals are computed.}

\item{level}{A numeric scalar between 0 and 1 giving the confidence level for
the intervals (if any) to be calculated. Used when \code{interval = TRUE}}

\item{...}{Additional arguments. Give \code{lag} and \code{horizon} if they are not
defined in \code{x} for the \code{empirical} model.}
}
\value{
A list of kriging forecasts (and prediction intervals).
}
\description{
Obtain kriging forecasts for an \code{mcgf_rs} object.
}
\details{
It produces simple kriging forecasts for a zero-mean mcgf. It supports
kriging for the \code{empirical} model, the \code{base} model, and the \code{all} model
which is the general stationary model with the base and Lagrangian model
from \code{x}.

When \code{soft = TRUE}, \code{prob} will be used to compute the soft forecasts
(weighted forecasts). The number of columns must match the number of unique
levels in \code{x}. The column order must be the same as the order of regimes as
in \code{levels(attr(x, "label", exact = TRUE))}. If not all regimes are seen in
\code{newlabel}, then only relevant columns in \code{prob} are used.

When \code{interval = TRUE}, confidence interval for each forecasts and each
horizon is given. Note that it does not compute confidence regions.
}
\seealso{
Other {functions related to prediction}: 
\code{\link{krige.mcgf}()},
\code{\link{krige}()}
}
\concept{{functions related to prediction}}
