% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krige.R
\name{krige}
\alias{krige}
\title{Generic function for displaying fitted models for mcgf objects}
\usage{
krige(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
Kriging results of \code{x}
}
\description{
Generic function for displaying fitted models for mcgf objects
}
\seealso{
Other {functions related to prediction}: 
\code{\link{krige.mcgf}()}
}
\concept{{functions related to prediction}}
