% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model.mcgf}
\alias{model.mcgf}
\title{Display fitted models for an mcgf object}
\usage{
\method{model}{mcgf}(
  x,
  model = c("all", "base", "lagrangian"),
  old = FALSE,
  print_model = TRUE,
  ...
)
}
\arguments{
\item{x}{An mcgf object.}

\item{model}{Which model to display.}

\item{old}{Logical; TRUE if the old model needs to be printed.}

\item{print_model}{Logical; TRUE if time lag and forecast horizon need to be
printed.}

\item{...}{Additional arguments. Not in use.}
}
\value{
None (invisible \code{NULL}).
}
\description{
Display fitted models for an mcgf object
}
\seealso{
Other {functions related to model fitting}: 
\code{\link{add_base.mcgf_rs}()},
\code{\link{add_base.mcgf}()},
\code{\link{add_lagr.mcgf}()},
\code{\link{add_lagr}()},
\code{\link{fit_base.mcgf_rs}()},
\code{\link{fit_base.mcgf}()},
\code{\link{fit_base}()},
\code{\link{fit_lagr.mcgf}()},
\code{\link{fit_lagr}()},
\code{\link{model}()}
}
\concept{{functions related to model fitting}}
