% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Generic function for displaying fitted models for mcgf objects}
\usage{
model(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
Fitted models for an mcgf object.
}
\description{
Generic function for displaying fitted models for mcgf objects
}
\seealso{
Other {functions related to model fitting}: 
\code{\link{add_base.mcgf_rs}()},
\code{\link{add_base.mcgf}()},
\code{\link{add_lagr.mcgf}()},
\code{\link{add_lagr}()},
\code{\link{fit_base.mcgf_rs}()},
\code{\link{fit_base.mcgf}()},
\code{\link{fit_base}()},
\code{\link{fit_lagr.mcgf}()},
\code{\link{fit_lagr}()},
\code{\link{model.mcgf}()}
}
\concept{{functions related to model fitting}}
