% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcgf.R
\name{new_mcgf}
\alias{new_mcgf}
\title{Create an mcgf object}
\usage{
new_mcgf(data, locations, dists, time)
}
\arguments{
\item{data}{Time series data set in space-wide format.}

\item{locations}{A matrix of data.frame of 2D points, first column
longitude, second column latitude, both in decimal degrees. Required when
\code{dists} is not supplied.}

\item{dists}{List of signed distance matrices. Required when \code{locations} is
not supplied.}

\item{time}{Optional, a vector of equally spaced time stamps.}
}
\value{
An S3 object of class \code{mcgf}. As it inherits and extends the
\code{data.frame} class, all methods remain valid to the \code{data} part of the
object. Additional attributes may be assigned and extracted.
}
\description{
Create an mcgf object
}
\keyword{internal}
