% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{obj_mle}
\alias{obj_mle}
\title{Title}
\usage{
obj_mle(par, cor_fn, x, lag, par_fixed)
}
\arguments{
\item{par}{Parameters of \code{cor_fn}.}

\item{cor_fn}{Correlation function}

\item{x}{An \code{mcgf} or \code{mcgf_rs} object}

\item{lag}{Time lag.}

\item{par_fixed}{Fixed parameters of \code{cor_fn}.}
}
\value{
The objective of maximum likelihood: the additive inverse of
log-likelihood.
}
\description{
Title
}
\keyword{internal}
