% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mcgf.R
\name{print.mcgf}
\alias{print.mcgf}
\title{Print an \code{mcgf} object.}
\usage{
\method{print}{mcgf}(x, attr = ".Data", ...)
}
\arguments{
\item{x}{An \code{mcgf} object.}

\item{attr}{Attribute to be printed.}

\item{...}{Optional arguments to print methods.}
}
\description{
Print an \code{mcgf} object.
}
\seealso{
Other {functions related to the class}: 
\code{\link{acfs.mcgf}()},
\code{\link{acfs}()},
\code{\link{add_acfs.mcgf}()},
\code{\link{add_acfs}()},
\code{\link{add_ccfs.mcgf}()},
\code{\link{add_ccfs}()},
\code{\link{ccfs.mcgf}()},
\code{\link{ccfs}()},
\code{\link{dists.mcgf}()},
\code{\link{dists}()},
\code{\link{fit_base}()},
\code{\link{mcgf}()},
\code{\link{sds}()}
}
\concept{{functions related to the class}}
