% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds.R
\name{sds.mcgf}
\alias{sds.mcgf}
\alias{sds.mcgf_rs}
\alias{sds<-}
\title{Extract, calculate, or assign standard deviations for an \code{mcgf} or
\code{mcgf_rs} object.}
\usage{
\method{sds}{mcgf}(x, ...)

\method{sds}{mcgf_rs}(x, ...)

sds(x) <- value
}
\arguments{
\item{x}{An \code{mcgf} or \code{mcgf_rs} object.}

\item{...}{Additional parameters or attributes. Not in use.}

\item{value}{A vector of standard deviations for all stations.}
}
\value{
\code{\link[=sds]{sds()}} returns empirical (regime-switching) standard deviations.
}
\description{
Extract, calculate, or assign standard deviations for an \code{mcgf} or
\code{mcgf_rs} object.
}
\details{
For \code{mcgf} objects, \code{\link[=sds]{sds()}} extracts or computes the empirical standard
deviations. The output is a vector of sds.

For \code{mcgf_rs} objects, \code{\link[=sds]{sds()}} extracts or computes the regime-switching
empirical standard deviations. The output is a list of vectors of sds. Each
element in the list corresponds to the sds for a regime.

\code{\link{sds<-}} assigns \code{sds} to \code{x}. Use \code{\link[=add_ccfs]{add_ccfs()}} to add both \code{ccfs} and
\code{sds} to \code{x}.
}
