% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sds.R
\name{sds}
\alias{sds}
\title{Generic function for standard deviations for each column}
\usage{
sds(x, ...)
}
\arguments{
\item{x}{An \strong{R} object.}

\item{...}{Additional parameters or attributes.}
}
\value{
Standard deviations.
}
\description{
Generic function for standard deviations for each column
}
\seealso{
Other {functions related to the class}: 
\code{\link{acfs.mcgf}()},
\code{\link{acfs}()},
\code{\link{add_acfs.mcgf}()},
\code{\link{add_acfs}()},
\code{\link{add_ccfs.mcgf}()},
\code{\link{add_ccfs}()},
\code{\link{ccfs.mcgf}()},
\code{\link{ccfs}()},
\code{\link{dists.mcgf}()},
\code{\link{dists}()},
\code{\link{fit_base}()},
\code{\link{mcgf}()},
\code{\link{print.mcgf}()}
}
\concept{{functions related to the class}}
