% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind.R
\docType{data}
\name{wind}
\alias{wind}
\alias{`wind_loc`}
\alias{wind_loc}
\title{Ireland wind data, 1961-1978}
\format{
\code{wind}: an object of class data.frame with 6574 rows and 12 columns.

\code{wind_loc}: an object of class data.frame with 12 rows and 2 columns.
}
\usage{
wind

wind_loc
}
\description{
Daily average wind speeds for 1961-1978 at 11 synoptic meteorological
stations in the Republic of Ireland (Haslett and raftery 1989). Wind speeds
are in m/s.
}
\details{
The data were obtained from the \strong{gstat} package, and were modified so that
the first column is the time stamps. Locations of the 11 stations are given
in \code{wind_loc}.
}
\references{
Haslett, J. and Raftery, A. E. (1989). Space-time Modelling with Long-memory
Dependence: Assessing Ireland's Wind Power Resource (with Discussion).
Applied Statistics 38, 1-50.

Gneiting, T., Genton, M., & Guttorp, P. (2006). Geostatistical Space-Time
Models, Stationarity, Separability, and Full Symmetry. In C&H/CRC
Monographs on Statistics & Applied Probability (pp. 151–175).
Chapman and Hall/CRC.
}
\keyword{datasets}
